+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### **origin/HEAD** 2020/09/01 cyan00@gmail.com
 
+- updated
 
 ### **v2.0.5** 2020/09/01 cyan00@gmail.com